Rename fallback component file
